Allow SIGUSR1 to reset feeds_stale health concern:
The rationale for this is that when whoever is monitoring the app is alerted of feeds_stale, they might investigate and find that the feeds are not in-fact 'stale', because there simply have not been any videos published for an extended period. In that case, it's useful to be able to 'disarm' the alert without having the restart the app.