Determine our own version (git SHA) using info embedded by the compiler
Fixes #42

Now requires Go 1.18+. With this, there is no longer any reason to use mage & the Magefile as a build script.

I stopped printing "build time", even although it is available (as "vcs.time"). That time is misleading in the case of a "dirty build" (modified or untracked files present) because it is the time of the commit that the dirty build is derived from, and not the time of the actual build.

Module related commands I ran:
    go mod edit -go 1.18
    go mod tidy